+++ flight/PiOS/inc/pios_flash.h
@@ -35,13 +35,13 @@ struct pios_flash_chunk {
 };
 
 struct pios_flash_driver {
-	int32_t (*start_transaction)(uint32_t flash_id);
-	int32_t (*end_transaction)(uint32_t flash_id);
-	int32_t (*erase_chip)(uint32_t flash_id);
-	int32_t (*erase_sector)(uint32_t flash_id, uint32_t addr);
-	int32_t (*write_data)(uint32_t flash_id, uint32_t addr, uint8_t * data, uint16_t len);
-	int32_t (*write_chunks)(uint32_t flash_id, uint32_t addr, struct pios_flash_chunk chunks[], uint32_t num_chunks);
-	int32_t (*read_data)(uint32_t flash_id, uint32_t addr, uint8_t * data, uint16_t len);
+	int32_t (*start_transaction)(uintptr_t flash_id);
+	int32_t (*end_transaction)(uintptr_t flash_id);
+	int32_t (*erase_chip)(uintptr_t flash_id);
+	int32_t (*erase_sector)(uintptr_t flash_id, uint32_t addr);
+	int32_t (*write_data)(uintptr_t flash_id, uint32_t addr, uint8_t * data, uint16_t len);
+	int32_t (*write_chunks)(uintptr_t flash_id, uint32_t addr, struct pios_flash_chunk chunks[], uint32_t num_chunks);
+	int32_t (*read_data)(uintptr_t flash_id, uint32_t addr, uint8_t * data, uint16_t len);
 };
 
 #endif	/* PIOS_FLASH_H_ */
